ScrappingWithPlaywright/MuseumDataScrapped/requirements.txt
playwright
pytest
pytest-asyncio
python-dotenv
shapely